+++ NEWS.md
@@ -2,6 +2,13 @@
 All notable changes to Seurat will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [Unreleased]
+
+### Added
+
+### Changes
+- Removed `add.iter` parameter from `RunTSNE` function
+
 ## [3.1.5] - 2020-04-14
 ### Added
 - New `scale` parameter in `DotPlot`